Simplify histogram style (#1373)
* Simplify histogram style

* Fix test

* Fix comments
