+++ src/app/search/search.component.scss
@@ -162,14 +162,6 @@ mat-tab-group.homeComponent {
   justify-content: space-between;
 }
 
-.workflow-histogram {
-  background-color: #d2fbf0;
-}
-
-.tool-histogram {
-  background-color: #d0effd;
-}
-
 mat-form-field {
   width: 100%;
 }
