+++ src/app/shared/pipe/pipe.module.ts
@@ -4,7 +4,7 @@ import { FilePathPipe } from '../../entry/file-path.pipe';
 import { TimeAgoMsgPipe } from '../../organizations/organization/time-ago-msg.pipe';
 import { GetFacetSearchUpdatePipe } from '../../search/facet-search/facet-search-update.pipe';
 import { GetFacetSearchResultsPipe } from '../../search/facet-search/facet-search.pipe';
-import { GetHistogramWidthPipe } from '../../search/get-histogram-width.pipe';
+import { GetHistogramStylePipe } from '../../search/get-histogram-style.pipe';
 import { MapFriendlyValuesPipe } from '../../search/map-friendly-values.pipe';
 import { SelectTabPipe } from '../entry/select-tab.pipe';
 
@@ -13,7 +13,7 @@ const DECLARATIONS: any[] = [
   MapFriendlyValuesPipe,
   SelectTabPipe,
   TimeAgoMsgPipe,
-  GetHistogramWidthPipe,
+  GetHistogramStylePipe,
   GetFacetSearchResultsPipe,
   GetFacetSearchUpdatePipe,
 ];
